# api/core.openapi.yaml
basePath: /
definitions:
...
  AllocationStatus:
    enum:
    - ACTIVE
    - INACTIVE
    - DELETED
    type: string
    x-enum-varnames:
    - ACTIVE
    - INACTIVE
    - DELETED
...
  ComputeAllocation:
    properties:
      compute_cluster_id:
        description: The ID of the compute cluster where the allocation is provisioned.
        type: string
      end_time:
        type: string
      id:
        type: string
      initial_su_amount:
        description: SUs allocated at the time of allocation creation.
        type: integer
      name:
        type: string
      posix_group:
        description: Fixed at creation; the group's name everywhere.
        type: string
      project_id:
        type: string
      start_time:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/AllocationStatus'
        description: ACTIVE, INACTIVE, DELETED, etc.
    type: object
...
  ComputeAllocationResource:
    properties:
      compute_cluster_id:
        description: The ID of the compute cluster the resource (partition) belongs
          to.
        type: string
      id:
        type: string
      name:
        description: resource / partition name, e.g., "cpu-01", "gpu-01", "gpu-interactive",
          etc.
        type: string
      resource_amount:
        description: Number of CPUs, GPUs.
        type: integer
      resource_type:
        description: 'TRES: cpu, gres/gpu'
        type: string
    type: object
  ComputeAllocationResourceMapping:
    properties:
      compute_allocation_id:
        type: string
      compute_allocation_resource_id:
        type: string
      id:
        type: string
      resource_amount:
        description: Amount of the resource allocated to this allocation (e.g., number
          of CPUs, GPUs).
        type: integer
      resource_time:
        description: Wall-clock time in minutes that the allocated amount is granted
          for.
        type: integer
    required:
    - compute_allocation_id
    - compute_allocation_resource_id
    - id
    - resource_amount
    - resource_time
    type: object
  ComputeAllocationResourceRate:
    properties:
      compute_allocation_resource_id:
        type: string
      end_time:
        description: The time when this rate expires.
        type: string
      id:
        type: string
      rate:
        description: The rate for the resource in SUs per unit, e.g., 0.5 SU per CPU
          hour, 2 SU per GPU hour, etc.
        type: number
      start_time:
        description: The time when this rate becomes effective.
        type: string
    type: object
  ComputeAllocationResourceSummary:
    properties:
      allocation_count:
        type: integer
      compute_cluster_id:
        description: The ID of the compute cluster the resource (partition) belongs
          to.
        type: string
      id:
        type: string
      name:
        description: resource / partition name, e.g., "cpu-01", "gpu-01", "gpu-interactive",
          etc.
        type: string
      rate_count:
        type: integer
      resource_amount:
        description: Number of CPUs, GPUs.
        type: integer
      resource_type:
        description: 'TRES: cpu, gres/gpu'
        type: string
      total_allocated:
        type: integer
      total_used_su:
        type: number
    required:
    - allocation_count
    - compute_cluster_id
    - id
    - name
    - rate_count
    - resource_amount
    - resource_type
    - total_allocated
    - total_used_su
    type: object
...
  attachResourceRequest:
    properties:
      compute_allocation_resource_id:
        type: string
      resource_amount:
        type: integer
      resource_time:
        type: integer
    type: object
...
  updateAllocationResourceMappingRequest:
    properties:
      resource_amount:
        type: integer
      resource_time:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 'REST API for Apache Custos: organizations, users, projects, compute
    clusters, allocations, audit traces, and the privilege/role authorization layer.
    OIDC bearer auth required. Connector endpoints live under /connectors/<name>/.'
  title: Apache Custos Core API
  version: 0.2.0
paths:
...
  /compute-allocation-resource-rates:
    post:
      consumes:
      - application/json
      description: Records the SU rate (e.g. SU per CPU-hour) for a resource over
        a time window.
      parameters:
      - description: Rate payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ComputeAllocationResourceRate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ComputeAllocationResourceRate'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a compute allocation resource rate
      tags:
      - Compute Allocation Resource Rates
  /compute-allocation-resource-rates/{id}:
    get:
      parameters:
      - description: Rate ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ComputeAllocationResourceRate'
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a compute allocation resource rate
      tags:
      - Compute Allocation Resource Rates
  /compute-allocation-resources:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ComputeAllocationResource'
            type: array
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all compute allocation resources
      tags:
      - Compute Allocation Resources
    post:
      consumes:
      - application/json
      description: Defines a resource (partition) — e.g. a CPU or GPU partition that
        allocations can be attached to.
      parameters:
      - description: Resource payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ComputeAllocationResource'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ComputeAllocationResource'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a compute allocation resource
      tags:
      - Compute Allocation Resources
  /compute-allocation-resources/{id}:
    get:
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ComputeAllocationResource'
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a compute allocation resource
      tags:
      - Compute Allocation Resources
  /compute-allocation-resources/{id}/allocations:
    get:
      parameters:
      - description: Compute allocation resource ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ComputeAllocation'
            type: array
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List compute allocations attached to a resource
      tags:
      - Compute Allocation Resources
...
  /compute-allocation-resources/{id}/rates:
    get:
      parameters:
      - description: Compute allocation resource ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ComputeAllocationResourceRate'
            type: array
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List rate history for a resource
      tags:
      - Compute Allocation Resource Rates
  /compute-allocation-resources/{id}/rates/effective:
    get:
      description: Returns the rate whose `[start_time, end_time)` window contains
        the `at` timestamp. Defaults to now when `at` is omitted.
      parameters:
      - description: Compute allocation resource ID
        in: path
        name: id
        required: true
        type: string
      - description: RFC3339 time; defaults to now
        in: query
        name: at
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ComputeAllocationResourceRate'
        "400":
          description: Invalid 'at' query parameter
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get the effective rate for a resource at a given time
      tags:
      - Compute Allocation Resource Rates
  /compute-allocation-resources/summary:
    get:
      description: Each row is a resource plus its allocation count, total allocated
        amount, total used SUs, and rate count.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ComputeAllocationResourceSummary'
            type: array
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List compute allocation resources with aggregate figures
      tags:
      - Compute Allocation Resources
...
  /compute-allocations/{id}/resources:
    get:
      description: Callers without the allocations read privilege must hold a membership
        on the allocation or a governance role on its project; others return 404.
      parameters:
      - description: Compute allocation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ComputeAllocationResource'
            type: array
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List resources attached to a compute allocation
      tags:
      - Compute Allocation Resources
    post:
      consumes:
      - application/json
      description: Creates a mapping between a compute allocation and a resource (partition)
        with a specific amount and wall-clock time grant.
      parameters:
      - description: Compute allocation ID
        in: path
        name: id
        required: true
        type: string
      - description: Attach payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/attachResourceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ComputeAllocationResourceMapping'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Attach a resource to a compute allocation
      tags:
      - Compute Allocation Resources
  /compute-allocations/{id}/resources/{resourceId}:
    delete:
      parameters:
      - description: Compute allocation ID
        in: path
        name: id
        required: true
        type: string
      - description: Compute allocation resource ID
        in: path
        name: resourceId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Detach a resource from a compute allocation
      tags:
      - Compute Allocation Resources
    get:
      description: Returns the amount and time granted to the allocation on the resource.
      parameters:
      - description: Compute allocation ID
        in: path
        name: id
        required: true
        type: string
      - description: Compute allocation resource ID
        in: path
        name: resourceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ComputeAllocationResourceMapping'
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a compute allocation -> resource mapping
      tags:
      - Compute Allocation Resources
    put:
      consumes:
      - application/json
      parameters:
      - description: Compute allocation ID
        in: path
        name: id
        required: true
        type: string
      - description: Compute allocation resource ID
        in: path
        name: resourceId
        required: true
        type: string
      - description: Mapping update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/updateAllocationResourceMappingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ComputeAllocationResourceMapping'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a compute allocation -> resource mapping
      tags:
      - Compute Allocation Resources
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"